refactor(pdf): extract relationship quality checks
